For those types that support empty make the to_foo return option
what happens if you use a label when the label is Label.empty. Before
this PR it would return a sentinal value. While this works it's not
functional, or even a good design in general. The options (wink) are
to raise an exception, which I think is bad, or to use option to avoid
borking the system. The second is what I've chosen.