enable deriving (#8)
* enable deriving

currently show an yojson with a hand written to_yojson and of_yojson
for vertex_set due the the handlableness of Set for
ppx_deriving_yojson.

* enhance vertex and graph tests

extend the tests for the vertex api and make sure to test both the
list and set versions of a graph. remove some left in printf's.