this seems like a much more productive representation
the original version of this was to much like the C++ version. while
the internals of the graph does need to be pretty mutable the API
wasn't at all what I'd like it to have been. It was both to complex
and to inflexible. This I think fixes most of those issues.