redo graph (#13)
* this seems like a much more productive representation

the original version of this was to much like the C++ version. while
the internals of the graph does need to be pretty mutable the API
wasn't at all what I'd like it to have been. It was both to complex
and to inflexible. This I think fixes most of those issues.

* update graph_api to match new graph

+ update (meaning comment out for the time being) graph_api
+ add a semigroup type
+ add a monoid type
+ reformat code again

semigroup and monoid are based mostly on the cats versions and are
targeted at allowing generic adaptation of the internal graph data
container and for graphs to be manipulated as a general container.