Moar/coverage (#16)
* create growable_array

* now with semigroup and monoid

and monoid "extends" or at least uses semigroup, at least for
definitional puposes. This also mutates the API for growable_array a
bit but I'm still not happy with it. graph is still moving so I'm
continuing to delay re-implementing the api module until I'm
satisified.

* add some typeclass like things

+ functor
+ applicative

* formatting

and a couple of concrete implementations

* add tests to the app

even though it's all stubs just to be sure that the setup of
bisect_ppx is correct.

* now with coverage we can expand again

this code has been stalled until it could be tested. Now that butter
case tests have been built laws based testing is the next frontier

* update and test growable_array

* reformat code

* Update README.md