fixes bug with graph cycles (#22)
* fixes bug with graph cycles

the check and add to the visited struct happened in the wrong place
for graphs that have cycles.

* add psq

it's not (yet) used but it will be needed for uniform cost
search (i.e. Dijkstra's) This shows the benefit of moving to Dune 1.10
lang and autogenerating the opam files rather than hand maintaining
them.